updating to add a static repos file
The static repos will allow you to add repositories to the packages.json by creating
a file in confs/ called static-repos.json. If this file exists, it will read through
it while generating the packages.json and append these repositories to the list. I've
added this so we can add certian things on github without having to define them in
every project that pulls in a project that requires one of these repos.